requirements.txt
streamlit>=1.28.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langgraph>=0.0.20
playwright>=1.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
fpdf2>=2.7.0
pyinstaller>=6.0.0


